refactor(testing): Move IteratingBytecode, FixedIterationsBytecode
