+++ packages/testing/src/execution_testing/__init__.py
@@ -90,7 +90,9 @@ from .tools import (
     Conditional,
     Create2PreimageLayout,
     DeploymentTestType,
+    FixedIterationsBytecode,
     Initcode,
+    IteratingBytecode,
     ParameterSet,
     Switch,
     While,
@@ -155,12 +157,14 @@ __all__ = (
     "EngineAPIError",
     "Environment",
     "EOA",
+    "FixedIterationsBytecode",
     "FixtureCollector",
     "Fork",
     "GasCosts",
     "Hash",
     "Header",
     "Initcode",
+    "IteratingBytecode",
     "JumpLoopGenerator",
     "Macro",
     "Macros",
