+++ packages/testing/src/execution_testing/tools/__init__.py
@@ -9,7 +9,9 @@ from .tools_code import (
     CodeGasMeasure,
     Conditional,
     Create2PreimageLayout,
+    FixedIterationsBytecode,
     Initcode,
+    IteratingBytecode,
     Switch,
     While,
 )
@@ -27,12 +29,14 @@ __all__ = (
     "Case",
     "CodeGasMeasure",
     "Conditional",
+    "Create2PreimageLayout",
     "DeploymentTestType",
+    "FixedIterationsBytecode",
     "Initcode",
+    "IteratingBytecode",
     "ParameterSet",
     "Switch",
     "While",
-    "Create2PreimageLayout",
     "extend_with_defaults",
     "gas_test",
     "generate_system_contract_deploy_test",
