+++ packages/testing/src/execution_testing/tools/tools_code/__init__.py
@@ -6,7 +6,9 @@ from .generators import (
     CodeGasMeasure,
     Conditional,
     Create2PreimageLayout,
+    FixedIterationsBytecode,
     Initcode,
+    IteratingBytecode,
     Switch,
     While,
 )
@@ -17,11 +19,13 @@ __all__ = (
     "Case",
     "CodeGasMeasure",
     "Conditional",
+    "Create2PreimageLayout",
+    "FixedIterationsBytecode",
     "Initcode",
+    "IteratingBytecode",
     "Solc",
     "Switch",
     "While",
     "Yul",
     "YulCompiler",
-    "Create2PreimageLayout",
 )
